+++ CHANGELOG.md
@@ -5,10 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v0.5.0](https://github.com/wowica/xogmios/releases/tag/v0.5.0) (2024-08-16)
 
-- Added partial support for Mempool monitoring mini-protocol. Allows reading transactions in the mempool.
-- Added generators mix task for generating boilerplate code for client modules: `mix help xogmios.gen.client`
+### Added
+
+- Initial support for Mempool monitoring mini-protocol. Allows reading transactions from the mempool.
+- mix task for generating boilerplate code for client modules. See `mix help xogmios.gen.client`
 
 ## [v0.4.1](https://github.com/wowica/xogmios/releases/tag/v0.4.1) (2024-06-05)